Add Swal to confirm the log out

+++ client/src/components/Header.jsx
@@ -2,17 +2,30 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { useAppState } from './AppContext'
 import { showToast } from '../toast'
+import Swal from 'sweetalert2'
 
 const Header = () => {
     const [state, dispatch] = useAppState()
 
     function logOut() {
-        localStorage.clear()
-        dispatch({
-            type: 'set-user',
-            payload: undefined
+        Swal.fire({
+            title: 'You are about to log out',
+            text: 'Are you sure?',
+            showCancelButton: true,
+            icon: 'warning',
+            cancelButtonText: 'No, go back',
+            confirmButtonText: 'Yes, log me out'
+        }).then(value => {
+            // console.log(value)
+            if (value.isConfirmed) {
+                showToast('info', 'You have logged out')
+                localStorage.clear()
+                dispatch({
+                    type: 'set-user',
+                    payload: undefined
+                })
+            }
         })
-        showToast('info', 'You have logged out')
 
     }
     return (
@@ -35,9 +48,9 @@ const Header = () => {
                         category: null,
                         id: ''}
                     }}>Add Note</Link>
-                    <Link onClick={logOut} to='/login'>
+                    <a href='#' onClick={logOut}>
                         Log Out
-                    </Link>
+                    </a>
                     </>
                     :
                     <Link to='/login' className='border-basic'>Log In</Link>
